code_training/requirements.txt
wheel
numpy
jax[cpu]
jaxtyping
dm-haiku
gymnax
wandb
optax
distrax
hydra-core
notebook
pandas
scipy
flashbax
tabulate
matplotlib
seaborn
pyyaml